+++ mayan/apps/common/dependencies.py
@@ -38,7 +38,7 @@ PythonDependency(
         ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
         (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
         SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='django', version_string='==2.1.15'
+    ''', module=__name__, name='django', version_string='==2.2.10'
 )
 PythonDependency(
     copyright_text='''
@@ -61,10 +61,10 @@ PythonDependency(
         LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
         OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
         SOFTWARE.
-    ''', module=__name__, name='PyYAML', version_string='==5.1.2'
+    ''', module=__name__, name='PyYAML', version_string='==5.3'
 )
 PythonDependency(
-    module=__name__, name='django-formtools', version_string='==2.1'
+    module=__name__, name='django-formtools', version_string='==2.2'
 )
 PythonDependency(
     copyright_text='''
@@ -98,7 +98,7 @@ PythonDependency(
     ''', module=__name__, name='django-pure-pagination', version_string='==0.3.0'
 )
 PythonDependency(
-    module=__name__, name='django-mathfilters', version_string='==0.4.0'
+    module=__name__, name='django-mathfilters', version_string='==1.0.0'
 )
 PythonDependency(
     copyright_text='''
@@ -130,7 +130,7 @@ PythonDependency(
         THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
         (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
         OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='django-model-utils', version_string='==3.1.2'
+    ''', module=__name__, name='django-model-utils', version_string='==4.0.0'
 )
 PythonDependency(
     copyright_text='''
@@ -155,35 +155,7 @@ PythonDependency(
         COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
         IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
         CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-    ''', module=__name__, name='django-mptt', version_string='==0.9.1'
-)
-PythonDependency(
-    copyright_text='''
-        Copyright (c) 2010, Matt Croydon, Mikhail Korobov
-        All rights reserved.
-
-        Redistribution and use in source and binary forms, with or without
-        modification, are permitted provided that the following conditions are met:
-        * Redistributions of source code must retain the above copyright
-        notice, this list of conditions and the following disclaimer.
-        * Redistributions in binary form must reproduce the above copyright
-        notice, this list of conditions and the following disclaimer in the
-        documentation and/or other materials provided with the distribution.
-        * Neither the name of the tastypie nor the
-        names of its contributors may be used to endorse or promote products
-        derived from this software without specific prior written permission.
-
-        THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-        ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-        WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-        DISCLAIMED. IN NO EVENT SHALL MATT CROYDON BE LIABLE FOR ANY
-        DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-        (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-        LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-        ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-        (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-        SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='django-qsstats-magic', version_string='==1.0.0'
+    ''', module=__name__, name='django-mptt', version_string='==0.11.0'
 )
 PythonDependency(
     module=__name__, name='django-stronghold', version_string='==0.4.0'
@@ -212,16 +184,16 @@ PythonDependency(
     ''', module=__name__, name='django-widget-tweaks', version_string='==1.4.5'
 )
 PythonDependency(
-    module=__name__, name='furl', version_string='==2.0.0'
+    module=__name__, name='furl', version_string='==2.1.0'
 )
 PythonDependency(
     module=__name__, name='gevent', version_string='==1.4.0'
 )
 PythonDependency(
-    module=__name__, name='gunicorn', version_string='==19.9.0'
+    module=__name__, name='gunicorn', version_string='==20.0.4'
 )
 PythonDependency(
-    module=__name__, name='mock', version_string='==2.0.0'
+    module=__name__, name='mock', version_string='==4.0.1'
 )
 PythonDependency(
     module=__name__, name='pathlib2', version_string='==2.3.5'
@@ -230,7 +202,7 @@ PythonDependency(
     copyright_text='''
         Author: Christian Theune
         License: LGPL 2.1
-    ''', module=__name__, name='pycountry', version_string='==18.12.8'
+    ''', module=__name__, name='pycountry', version_string='==19.8.18'
 )
 PythonDependency(
     copyright_text='''
@@ -253,10 +225,10 @@ PythonDependency(
         LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
         FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
         DEALINGS IN THE SOFTWARE.
-    ''', module=__name__, name='pytz', version_string='==2019.1'
+    ''', module=__name__, name='pytz', version_string='==2019.3'
 )
 PythonDependency(
-    module=__name__, name='requests', version_string='==2.21.0'
+    module=__name__, name='requests', version_string='==2.22.0'
 )
 PythonDependency(
     copyright_text='''
@@ -289,19 +261,19 @@ PythonDependency(
 
 PythonDependency(
     module=__name__, environment=environment_development, name='Werkzeug',
-    version_string='==0.15.4'
+    version_string='==1.0.0'
 )
 PythonDependency(
     module=__name__, environment=environment_development, name='devpi-server',
-    version_string='==5.3.1'
+    version_string='==5.4.0'
 )
 PythonDependency(
     environment=environment_development, module=__name__,
-    name='django-debug-toolbar', version_string='==1.11'
+    name='django-debug-toolbar', version_string='==2.2'
 )
 PythonDependency(
     environment=environment_development, module=__name__,
-    name='django-extensions', version_string='==2.1.9'
+    name='django-extensions', version_string='==2.2.8'
 )
 PythonDependency(
     environment=environment_development, help_text=_(
@@ -316,7 +288,7 @@ PythonDependency(
 PythonDependency(
     environment=environment_development, help_text=_(
         'Command line environment with autocompletion.'
-    ), module=__name__, name='ipython', version_string='==5.8.0'
+    ), module=__name__, name='ipython', version_string='==7.12.0'
 )
 PythonDependency(
     environment=environment_development, help_text=_(
@@ -329,7 +301,7 @@ PythonDependency(
 )
 PythonDependency(
     environment=environment_development,
-    module=__name__, name='transifex-client', version_string='==0.13.6'
+    module=__name__, name='transifex-client', version_string='==0.13.7'
 )
 
 # Testing
@@ -345,11 +317,11 @@ PythonDependency(
 )
 PythonDependency(
     environment=environment_testing, module=__name__, name='coverage',
-    version_string='==4.4.1'
+    version_string='==5.0.3'
 )
 PythonDependency(
     environment=environment_testing, module=__name__, name='coveralls',
-    version_string='==1.3.0'
+    version_string='==1.11.1'
 )
 PythonDependency(
     environment=environment_testing, module=__name__,
@@ -366,29 +338,29 @@ PythonDependency(
 )
 PythonDependency(
     environment=environment_testing, module=__name__, name='tox',
-    version_string='==3.8.6'
+    version_string='==3.14.5'
 )
 PythonDependency(
     environment=environment_testing, module=__name__, name='psutil',
-    version_string='==5.6.3'
+    version_string='==5.6.7'
 )
 
 # Build
 
 PythonDependency(
     environment=environment_build, module=__name__, name='twine',
-    version_string='==1.9.1'
+    version_string='==3.1.1'
 )
 PythonDependency(
     environment=environment_build, module=__name__, name='wheel',
-    version_string='==0.30.0'
+    version_string='==0.34.2'
 )
 
 # Documentation
 
 PythonDependency(
     environment=environment_documentation, module=__name__, name='Sphinx',
-    version_string='==1.8.5'
+    version_string='==2.4.1'
 )
 PythonDependency(
     environment=environment_documentation, module=__name__,
@@ -398,7 +370,7 @@ PythonDependency(
 PythonDependency(
     environment=environment_documentation, module=__name__,
     name='sphinx-sitemap',
-    version_string='==1.0.2'
+    version_string='==1.1.0'
 )
 PythonDependency(
     environment=environment_documentation, module=__name__,
@@ -408,16 +380,16 @@ PythonDependency(
 PythonDependency(
     environment=environment_documentation, module=__name__,
     name='sphinxcontrib-blockdiag',
-    version_string='==1.5.5'
+    version_string='==2.0.0'
 )
 PythonDependency(
     environment=environment_documentation, module=__name__,
     name='sphinxcontrib-spelling',
-    version_string='==4.2.1'
+    version_string='==4.3.0'
 )
 # sphinx-autobuild has a dependency on Tornado,
 # but Tornado 6.0 dropped support for Python 2.7
 PythonDependency(
     environment=environment_documentation, module=__name__, name='tornado',
-    version_string='<6.0'
+    version_string='==6.0.3'
 )