+++ mayan/apps/dependencies/dependencies.py
@@ -3,5 +3,5 @@ from __future__ import absolute_import, unicode_literals
 from .classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='node-semver', version_string='==0.6.1'
+    module=__name__, name='node-semver', version_string='==0.8.0'
 )