+++ HISTORY.rst
@@ -7,3 +7,6 @@
 - Disable constraing checking for SQLite.
 - Expose Django's ``AUTHENTICATION_BACKENDS`` setting.
 - Add support for list modes.
+- Update dependencies to the most recent version.
+- Move django-qsstats-magic to the mayan_statistics app.
+- Remove codecov from the dependencies.